Go back to double-reviewing PKGBUILDs
It's a nice idea to not have to look at the PKGBUILD twice, but the
edits that will have effect in the building would be the /second/ edit
so it needs to be there.

Users that don't want to see it twice can use `prompt` and say `n` to
the second review.

I'm still open to a "solution" but for now, this is better.