Gracefully handle missing targets
Aurget often finds a PKGBUILD for a package not actually in the AUR.
This can cause some confusing errors for dependency packages. These
cases typically indicate a package that has moved to the official
repositories but hasn't had its PKGBUILD cleaned off the AUR servers
yet.

In the case of a dependency, we should just skip it. Makepkg will either
find it or error more informatively itself. For normal arguments, we
continue to error; it's still invalid usage.