Don't use --pkg option with makepkg
This argument was removed from makepkg with the release of Pacman 5.0.
When a PKGBUILD containing a split package is built, all packages
contained therein are built.

Links:

- https://lists.archlinux.org/pipermail/pacman-dev/2015-September/020353.html
- https://projects.archlinux.org/pacman.git/commit/scripts/makepkg.sh.in?id=169287e494a5348687260a00697be06b36ba4434

Note: when --install is used, all packages are installed. Users that
don't want this are advised to build with aurget -Sb and install what
they want with pacman -U directly.

We could potentially simplify further, but removing this option should
fix the current bug of aurget not working at all with the latest Pacman
version.