Use stable order for available-upgrade names
Bash associative-array ordering is dependent on many things, and will
fail the upgrade test if the curl arguments are in a different order and
no fixture is found. By tracking the names in their own array, we
preserve the order that came out of pacman -Qm. This is a slightly
better UI but, more importantly, makes the test behavior consistent.